move FakeLazyObservation from tests/helper.py to tests/mapmaking/helper.py
